NET-17: move verify signature to proof
